requirements.txt
django==4.2
mysqlclient
Pillow
django-admin-interface
django-crispy-forms
crispy-bootstrap4
django-paypal
python-decouple
gunicorn